Sour 94 (#14)
* build: Upgrade boost to 1.83

* src: Replace logging macro with constexpr

* build: add library cmake function

* src: Create FilePath target

* merge with add cmake lib

* src: Create logging cmake target

* build: Add cmake format script

* src: Add utilityString&TextCodec targets

* build: Format add_sourcetrail_library

* src: Add ConfigManager target

* src: Add TimeStamp&FileInfo as targets

* src: Add TestCodec unit-test

* src: Add test cmake

* src: Add Version target

* src: Add utilityUuid target

* src: Add utility target

* src: Remove utilityWindow and library

* src: Add UnorderedCache target

* src: Create GlobalId and Tree target

* src: Remove trace

* src: Add Status target

* src: Move every component in dir

* src: Create TextAccess target

* src: Remove plugin for now

* src: Move commandline to core/utility

* src: Move file to core

* src: Formating

---------

+++ src/lib/component/controller/StatusController.cpp
@@ -1,5 +1,5 @@
 #include "StatusController.h"
-// internal
+
 #include "IApplicationSettings.hpp"
 #include "StatusView.h"
 #include "utility.h"
@@ -56,7 +56,7 @@ void StatusController::addStatus(const std::vector<Status>& statuses) {
   std::vector<Status> filteredStatus;
 
   for(const Status& status : statuses) {
-    if((status.type & m_statusFilter) != 0) {
+    if((static_cast<std::uint8_t>(status.type) & static_cast<std::uint8_t>(m_statusFilter)) != 0) {
       filteredStatus.push_back(status);
     }
   }
